VoynichBackend/config/celery_app.py
from __future__ import absolute_import
from configurations import importer
from celery import Celery
import os

os.environ.setdefault("DJANGO_SETTINGS_MODULE", "config.settings.development")

importer.install()

app = Celery('config')

CELERY_TIMEZONE = 'UTC'

app.config_from_object('django.conf:settings')
app.autodiscover_tasks()
