VoynichBackend/requirements.txt
django-configurations
djangorestframework
django-cors-headers
psycopg2-binary
django-haystack
dj-database-url
django-nose
django-material
django-viewflow
django-storages
elasticsearch==2.0.0
cryptography
gunicorn
channels
aioredis
asgiref
coreapi
celery
boto3
Django
isort
nose
PyJWT
